Started impementation of Workbench system
Begun the workbench implementation for RatRunner. Currently the
ScriptableObject exists and I am attempting to design the system in
a way that allows for minimal code later on when adding
any new elements to the game such as more recipes.

As it stands, the workbench only has some public fields
and no functionality. While upgrade manager has some basic
functionality to add upgrades, there is no way to actually apply them
in game yet since there is no place to apply them, i.e a player script.

Next steps will be to finish the
workbench functionality (UI and interaction),
add missing features to upgrade manager
(such as removing used items from inventory),
and adapt the codebase to work with the player script. I will also need
to create some basic recipes to test the system and revise the inventory
and item scripts to ensure compatibility
 with the upgrade/workbench system.

The main issue I foresee is ensuring that items used in upgrades
are properly removed from the inventory and that the upgrades
are correctly applied to the player stats or abilities. Alongside this,
the saving system will need to be updated to account for upgrades and
the new inventory format(its now a dictionary).